fix backend/predict-api/requirements.txt
fastapi
tensorflow
numpy
google-cloud-firestore
uvicorn
Pillow
